backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow==10.1.0
python-dotenv==1.0.0
stripe==7.5.0
django-storages==1.14.2
django-filter==23.3
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
django-cleanup==8.0.0
django-environ==0.11.2
psycopg2-binary==2.9.9
dj-database-url==2.1.0 